fix: refactorizado de nuevo
